+++ app/src/main/java/io/bloco/circles/ui/shared/MessageLoader.kt
@@ -20,6 +20,14 @@ class MessageLoader
 
     fun show(string: String) =
         Snackbar.make(rootView, string, Snackbar.LENGTH_LONG)
-        .show()
+            .show()
+
+    fun show(@StringRes message: Int, @StringRes action: Int, listener: View.OnClickListener,) =
+        Snackbar.make(rootView, message, Snackbar.LENGTH_LONG)
+            .setAction(action, listener).show()
+
+    fun show(message: String, action: String, listener: View.OnClickListener, ) =
+        Snackbar.make(rootView, message, Snackbar.LENGTH_LONG)
+            .setAction(action, listener).show()
 
 }